改完文章和購物車 自己測試時沒問題 看其他人測試有沒有其他狀況 然後管理員目前無法刪除大叔 有遇到問題是因為文章有外鍵 所以無法刪除 有事過加cascade 但是沒有效果

+++ ossanRental/src/main/java/_03_listOssans/controller/BuyOssanServlet.java
@@ -37,7 +37,7 @@ public class BuyOssanServlet extends HttpServlet {
 		String name 		= request.getParameter("name");
 		String nickname 	= request.getParameter("nickname");
 		String qtyStr 		= request.getParameter("qty");
-		Double price;
+		int price;
 		Double discount;
 		int qty = 0 ; 
 		int pKey = 0;
@@ -50,7 +50,7 @@ public class BuyOssanServlet extends HttpServlet {
 			throw new ServletException(e); 
 		}		
 		//暫時放一筆固定價格, Jay 20190503
-		price = (double) 300;
+		price = 300;
 		discount = (double) 1;		
 		// 將訂單資料封裝到OrderItem物件內
 		OrderItem oi = new OrderItem(pKey,email,name,nickname,qty,price,discount);
